Added request collab functionality
Under BrandViewCC.vue, new function, 'requestCollab' added. Calls the complex place_request microservice.

Modified the URL part in place_request.py  as it was pointing to the wrong URL.

I believe there was also a wrong declaration of the externalService in the yaml file, so changed it accordingly.

Changed back the IP addresses so there's no need to manually change back @Repurica